requirements.txt
fastapi
starlette
uvicorn
authlib
python-dotenv
SQLAlchemy
alembic
markdown
beautifulsoup4
asgiref
pyarrow
